requirements.txt
langchain
langchain-community
langchain-huggingface
faiss-cpu
pypdf
huggingface-hub
flasK
python-dotenv
sentence-transformers
